requirements.txt
asyncio==3.4.3
aiohttp==3.12.13
aiosqlite==0.21.0
pydantic==2.11.7
beautifulsoup4==4.13.4
python-dotenv==1.2.1
lxml==6.0.0
pytest==8.4.1
pytest-asyncio==1.0.0
beautifulsoup4~=4.13.4
